deploy_docker/requirements.txt
chromadb==0.5.3
ffmpeg-python==0.2.0
flask==3.0.3
ipython==8.15.0
ipywidgets==8.1.3
langchain==0.2.7
langchain-community==0.2.6
langchain-openai==0.1.16
langgraph==0.1.8
moviepy==1.0.3
openai-whisper==20231117
pyannote.audio==3.3.1
python-dotenv==1.0.1
torch==2.3.1
whisper==1.1.10
yt-dlp==2024.7.9